Fix for startup issue crash in 5.1.1 in Home.java
This fix is a proof of concept only.

There is a race condition where I don’t think the DB services are up
yet in 5.1.1 but the Launcher will already be running. There are
multiple calls to the requestAllPermission() which fires off the
openDBandLoadApp() function. With the DB services not up yet, it
crashes the app.

Waiting for 10 seconds for Android to boot up and start the DB services
solves this problem.

There is DEFINITELY a better solution to this.